# api/openapi.yaml
openapi: 3.0.3
info:
  title: EgressDeck Controller HTTP API
  version: 0.1.0
  description: Current Go HTTP contract. Snake_case wire values are mapped by the Vue client into UI summaries.
    Collection pages use items, limit, total and optional next_cursor; the durable event feed uses its own resume
    cursor. Conditional remote adapter routes return typed HTTP 501 when unavailable. A 202 identifies an existing
    controller journal entry and is separate from applied, observed and traffic-verified state. Roadmap-only operations
    are outside paths under x-planned-operations; OIDC callback registration follows exact OIDC_REDIRECT_URL under
    x-oidc-callback.
servers:
- url: /
security:
- sessionCookie: []
...
paths:
...
  /api/v1/providers/{id}/revisions/{revision}/apply:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    - name: revision
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    post:
      tags:
      - providers
      summary: Publish and read back a provider revision
      operationId: applyProviderRevision
      x-required-capability: provider.publish_hot
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/RuntimeIfMatch'
      responses:
        '202':
          description: Journal operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAccepted'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '501':
          $ref: '#/components/responses/NotImplemented'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '412':
          description: Expected selection/provider revision differs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '428':
          description: If-Match is required; use zero for the first mutation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid selection scope or other typed input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      x-required-role: operator
...
  /api/v1/outbound-groups:
    get:
      tags:
      - policy
      summary: List outbound groups
      operationId: listOutboundGroups
      responses:
        '200':
          description: Groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundGroupList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: invalid_cursor or invalid query/limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'stale_cursor: snapshot changed or cursor expired'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: collection_item_too_large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: collection_capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Returns a bounded items page with limit, total and optional next_cursor. Default limit 100, maximum
        500. Follow next_cursor using the same query and limit. Snapshot change or expiry returns 409; restart invalidates
        the cursor. Responses are bounded to 4 MiB, scan to 100000 items/64 MiB; live observation values can refresh
        between pages.
      parameters:
      - $ref: '#/components/parameters/CollectionLimit'
      - $ref: '#/components/parameters/CollectionCursor'
      x-required-role: viewer
    post:
      tags:
      - policy
      summary: Create outbound group
      operationId: createOutboundGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutboundGroupCreate'
      responses:
        '201':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundGroup'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      x-required-role: admin
  /api/v1/outbound-groups/{id}:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    get:
      tags:
      - policy
      summary: Read outbound group
      operationId: getOutboundGroup
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundGroup'
        '404':
          $ref: '#/components/responses/Error404'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      x-required-role: viewer
    put:
      tags:
      - policy
      summary: Replace outbound group
      operationId: replaceOutboundGroup
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/Revision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutboundGroupCreate'
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundGroup'
        '409':
          description: outcome_unknown; an earlier group mutation must be reconciled before editing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/responses/Error412'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '428':
          description: Expected revision header/query required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Replaces the entire resource; PATCH is not JSON merge patch. Supply If-Match or revision query.
        Omitted mutable fields use zero values. This updates desired configuration only; use the separate apply operation to deploy it against the active provider inventory.
      x-required-role: admin
    patch:
      tags:
      - policy
      summary: Patch outbound group
      operationId: patchOutboundGroup
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/Revision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutboundGroupCreate'
      responses:
        '200':
          description: Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundGroup'
        '409':
          description: outcome_unknown; an earlier group mutation must be reconciled before editing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/responses/Error412'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '428':
          description: Expected revision header/query required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Replaces the entire resource; PATCH is not JSON merge patch. Supply If-Match or revision query.
        Omitted mutable fields use zero values. This updates desired configuration only; use the separate apply operation to deploy it against the active provider inventory.
      x-required-role: admin
  /api/v1/outbound-groups/{id}/apply:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    post:
      tags:
      - policy
      summary: Publish desired outbound-group configuration independently of provider content
      operationId: applyOutboundGroup
      x-required-capability: group.publish_hot
      x-required-role: admin
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/IdempotencyKey'
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
        description: Exact desired group revision to publish. Missing or invalid input returns 428; mismatch returns 412.
      description: Applies the stored desired group configuration against the active provider inventory without refetching or advancing its connection revision. Requires configured group publication/readback adapters and explicit native group bindings. The idempotency key binds the full normalized desired group configuration. Applied and observed fields advance only after matching readback. Definite preflight rejection records failed; an uncertain remote outcome records outcome_unknown and requires reconciliation.
      responses:
        '202':
          description: Durable group operation; group is present after matching applied readback and omitted on idempotent replay or uncertain outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupApplyAccepted'
        '400':
          description: Invalid runtime publication request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          description: Idempotency conflict, unresolved prior operation, provider/group mismatch or definite native rejection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/responses/Error412'
        '422':
          $ref: '#/components/responses/Error422'
        '428':
          description: If-Match with the desired group revision is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          $ref: '#/components/responses/NotImplemented'
        '503':
          $ref: '#/components/responses/Error503'
...
  /api/v1/gateways/{id}/probes/groups/{groupId}:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    - name: groupId
      in: path
      required: true
      schema:
        type: string
    post:
      summary: probeGatewayGroup
      operationId: probeGatewayGroup
      tags:
      - gateways
      x-implementation-status: implemented
      x-required-capability: probe.group
      description: Checks live gateway capabilities on every request. Query parameters and browser-selected probe
        URLs are rejected. Results are gateway-origin observations, not client traffic verification.
      responses:
        '200':
          description: Gateway diagnostic observation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayProbeResponse'
        '400':
          description: Query parameters or nonempty probe body rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gateway or scoped target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Gateway diagnostic capability is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsupportedResponse'
        '502':
          description: invalid_gateway_response or gateway_unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      parameters:
      - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties: {}
      x-required-role: operator
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: egressdeck_session
      description: HTTP-only session; mutations also require X-CSRF-Token matching egressdeck_csrf.
...
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: Expected positive resource revision, optionally quoted. Current mutating handlers require this
        header or the revision query; missing or invalid input returns 428.
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Optional controller journal key. Reuse only for the same action/payload. No HTTP-layer replay
        guarantee is provided by the gateway agent.
    Revision:
      name: revision
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
      description: Alternative to If-Match; one positive expected revision is required.
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: Required with a session in session/header mode. Must match egressdeck_csrf cookie and session
        token.
    RuntimeIfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: Expected nonnegative selection/active-provider revision, optionally quoted. Use 0 for first selection/publication.
        Missing or invalid header returns 428; no query fallback.
    CollectionLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    CollectionCursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 2048
      description: Opaque next_cursor from prior page. Bound to route, query, limit and stable snapshot. Restart
        invalidates cursors; expiry is 15 minutes.
  responses:
    NotImplemented:
      description: Adapter capability is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnsupportedResponse'
    Error401:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: Forbidden or CSRF failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error412:
      description: Revision conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error422:
      description: Semantic validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error503:
      description: Dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorDetail:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        capability:
          type: string
    ErrorResponse:
      oneOf:
      - type: object
        required:
        - error
        properties:
          error:
            $ref: '#/components/schemas/ErrorDetail'
      - type: object
        required:
        - error
        - message
        properties:
          error:
            type: string
          message:
            type: string
    UnsupportedResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
              - unsupported_capability
            message:
              type: string
            capability:
              type: string
          additionalProperties: false
          required:
          - code
          - message
          - capability
      additionalProperties: false
      required:
      - error
...
    ProviderRevision:
      type: object
      properties:
        provider_id:
          type: string
        number:
          type: integer
          format: int64
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
          nullable: true
        parse_report:
          $ref: '#/components/schemas/ParseReport'
        changes:
          $ref: '#/components/schemas/ChangeReport'
        state:
          type: string
        created_at:
          type: string
          format: date-time
        published_at:
          nullable: true
          type: string
          format: date-time
      additionalProperties: false
      required:
      - provider_id
      - number
      - nodes
      - parse_report
      - changes
      - state
      - created_at
      x-go-type: providers.Revision
      x-source: internal/providers/lifecycle.go
      description: Public revision JSON deliberately excludes private provider content hashes. Nodes exclude identity/credential
        hashes and credentials.
...
    OutboundGroupCreate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        gateway_id:
          type: string
        node_ids:
          type: array
          items:
            type: string
          nullable: true
          minItems: 1
        mode:
          type: string
          enum:
          - manual
        replacement_policy:
          type: string
          enum:
          - block
          - none
        source_filters:
          $ref: '#/components/schemas/SourceFilters'
      additionalProperties: false
      required:
      - name
      - gateway_id
      - node_ids
      description: gateway_id must exist and node_ids must name at least one supported staged/active node. Automatic
        selection currently returns 422 until a qualified chooser exists.
    OutboundGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        gateway_id:
          type: string
        node_ids:
          type: array
          items:
            type: string
          nullable: true
        mode:
          type: string
          enum:
          - manual
          - automatic
        replacement_policy:
          type: string
          enum:
          - block
          - none
        revision:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        source_filters:
          $ref: '#/components/schemas/SourceFilters'
        selection_scope:
          type: string
          enum:
          - shared_tcp_udp
          - independent_transport
          description: Runtime selection semantics for this group. Shared selection uses one TCP/UDP control and requires both scopes in one mutation.
        applied_revision:
          type: integer
          format: int64
          minimum: 0
          description: Last independently applied group configuration revision; omitted until publication is observed.
        observed_revision:
          type: integer
          format: int64
          minimum: 0
        applied_node_ids:
          type: array
          items:
            type: string
        observed_node_ids:
          type: array
          items:
            type: string
        applied_generation:
          type: integer
          format: int64
          minimum: 0
        observed_generation:
          type: integer
          format: int64
          minimum: 0
      additionalProperties: false
      required:
      - id
      - name
      - gateway_id
      - node_ids
      - mode
      - replacement_policy
      - revision
      - created_at
      - updated_at
      - selection_scope
      x-go-type: api.outboundGroupView
      x-source: internal/api/services.go
      description: revision and node_ids are desired group configuration. Applied and observed fields advance only after matching gateway readback; editing membership does not advance a provider connection revision or automatically deploy the group.
...
    Selection:
      type: object
      properties:
        group_id:
          type: string
        scope:
          $ref: '#/components/schemas/OutboundsScope'
        desired_node_id:
          type: string
        applied_node_id:
          type: string
        observed_node_id:
          type: string
        generation:
          type: integer
          format: int64
        applied_generation:
          type: integer
          format: int64
        observed_generation:
          type: integer
          format: int64
        unavailable:
          type: boolean
        revision:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - group_id
      - scope
      - generation
      - applied_generation
      - observed_generation
      - unavailable
      - revision
      - updated_at
      x-go-type: outbounds.Selection
      x-source: internal/outbounds/outbounds.go
...
    GroupApplyAccepted:
      type: object
      required:
      - operation_id
      - status
      properties:
        operation_id:
          type: string
        status:
          type: string
          description: Durable operation status; applied confirms matching management readback, not traffic verification.
        group:
          $ref: '#/components/schemas/OutboundGroup'
      additionalProperties: false
...
    OperationAccepted:
      type: object
      properties:
        operation_id:
          type: string
        status:
          type: string
          enum:
          - draft
          - validated
          - staged
          - applying
          - verifying
          - applied
          - partially_applied
          - failed
          - outcome_unknown
        revision:
          $ref: '#/components/schemas/ProviderRevision'
        changes:
          $ref: '#/components/schemas/ChangeReport'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Selection'
          nullable: true
      additionalProperties: false
      required:
      - operation_id
      - status
      description: References an existing journal entry at /api/v1/operations/{id}. Optional provider/selection
        payloads may be omitted on replay. HTTP 202 can contain a settled status; read the operation instead of
        assuming asynchronous execution or traffic success.
...
    OutboundGroupList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OutboundGroup'
        limit:
          type: integer
          minimum: 1
          maximum: 500
        total:
          type: integer
          minimum: 0
        next_cursor:
          type: string
      additionalProperties: false
      required:
      - items
      - limit
      - total
...
    ChangeReport:
      type: object
      properties:
        added:
          type: array
          items:
            $ref: '#/components/schemas/ProvidersNodeChange'
          nullable: true
        removed:
          type: array
          items:
            $ref: '#/components/schemas/ProvidersNodeChange'
          nullable: true
        changed:
          type: array
          items:
            $ref: '#/components/schemas/ProvidersNodeChange'
          nullable: true
        renamed:
          type: array
          items:
            $ref: '#/components/schemas/ProvidersNodeChange'
          nullable: true
        ambiguous:
          type: array
          items:
            $ref: '#/components/schemas/ProvidersIdentityAmbiguity'
          nullable: true
        unchanged:
          type: integer
          format: int64
        noop:
          type: boolean
      additionalProperties: false
      required:
      - unchanged
      - noop
      x-go-type: providers.ChangeReport
      x-source: internal/providers/lifecycle.go
...
    GatewayProbeResponse:
      type: object
      additionalProperties: false
      properties:
        gateway_id:
          type: string
        gateway_name:
          type: string
        observed_at:
          type: string
          format: date-time
          description: Controller receipt time.
        target_kind:
          type: string
          enum:
          - node
          - group
        target_id:
          type: string
        probe:
          $ref: '#/components/schemas/DiagnosticProbe'
      required:
      - gateway_id
      - gateway_name
      - observed_at
      - target_kind
      - target_id
      - probe
...
    SourceFilters:
      type: object
      additionalProperties: false
      properties:
        provider_ids:
          type: array
          items:
            type: string
          maxItems: 1024
        protocols:
          type: array
          items:
            type: string
          maxItems: 1024
        exclude_provider_ids:
          type: array
          items:
            type: string
          maxItems: 1024
        exclude_protocols:
          type: array
          items:
            type: string
          maxItems: 1024
        include_node_ids:
          type: array
          items:
            type: string
          maxItems: 1024
        exclude_node_ids:
          type: array
          items:
            type: string
          maxItems: 1024
        include_names:
          type: array
          items:
            type: string
          maxItems: 1024
        exclude_names:
          type: array
          items:
            type: string
          maxItems: 1024
      description: At least one positive constraint is required. Provider/protocol restrictions are ANDed; included
        IDs OR included exact case-sensitive names can match. Exclusions always win. Names are not regular expressions.
      x-go-type: outbounds.SourceFilters
      x-source: internal/outbounds/filters.go
...